Refactor to make it more testable
